add validation when delete supplier and category, on remove we alert the relation that category and supplier has with product

+++ src/domain/use-cases/category-remove.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ConflictException,
+} from '@nestjs/common';
 
 import { CategoryRepository } from '@domain/repositories/category.repository';
 
@@ -19,6 +23,16 @@ export class CategoryRemove {
 
     if (!categoryToRemove) {
       throw new NotFoundException();
+    }
+
+    const productsCount = await this.categoryRepository.countProducts(
+      categoryId,
+    );
+
+    if (productsCount > 0) {
+      throw new ConflictException(
+        `Category has relation with ${productsCount} products, remove the products to remove category`,
+      );
     }
 
     await this.categoryRepository.remove(categoryId);
